requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
boto3==1.9.134
botocore==1.12.134
bs4==0.0.1
cffi==1.12.3
constantly==15.1.0
contractions==0.0.18
cryptography==2.6.1
cssselect==1.0.3
docutils==0.14
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lxml==4.3.3
Markdown==3.1
mock==2.0.0
numpy==1.16.3
pandas==0.24.2
parsel==1.5.1
pbr==5.1.3
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
queuelib==1.5.0
s3transfer==0.2.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.4.0
termcolor==1.1.0
Twisted==18.9.0
urllib3==1.24.2
w3lib==1.20.0
Werkzeug==0.15.2
zope.interface==4.6.0
